refactor: move disk save helpers to service
## Centralize file save helpers
- wire fileSyncService with folder resolution callbacks
- relocate note and transcript save logic into service layer

## Other changes
- fix duplicate error log formatting
